Refactored
Renamed several files and functions for readability.
HTTP request now returns classes Timetable in json rather than starting a new scrapping process.
Added code for cancelling request, But commented it because it is an absurd approach.
Also The timetables that are scrapped by cronjob are stored in Timetable.json file in Outputs directory.